fix(debugger): crash when stepping through locations spanning multiple lines (#3920)
# Description

The debugger REPL was failing under certain scenarios when a stepped
location spanned multiple lines. We tracked down the failure to a
missing bounds check.

This PR fixes the issue, introduces a refactor to make the relevant code
testable, and adds regression tests to prevent reoccurrence.

This PR is best reviewed as a whole instead of commit-by-commit.

## Summary

- Moves source code printing logic from the main debugger REPL module,
to a new source_code_printer module.
- Decouples actual printing from rendering logic to improve testability.
- Adds regression tests

## Documentation

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.