chore(debugger): Integration tests (#3938)
# Description

Adds integration tests for the debugger. 

## Problem

Related to #3015. Generates a new suite of tests that uses the examples
at test_programs/execution_success to verify that the debugger can be
started and stepped to completion with those examples.

## Summary

We take the same approach used for the `nargo execute` cmd: generating
test cases from those at test_programs/execution_success. This means
future extensions to the Noir language captured in those examples will
automatically be included in the debugger integration suite, and will
help detect any potential divergence between the language runtime and
the debugger instrumentation.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.