feat: remove range constraints from witnesses which are constrained to be constants (#3928)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

A range constraint is a weaker constraint than directly asserting the
witness to be equal to a value (as long as that value fits within the
set number of bits). We can then remove range constraints from these
witnesses.

I've made sure to keep any range constraints which would conflict with
the constant witness assignment.

## Additional Context

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.