feat(lsp): cache definitions for goto requests (#3930)
# Description

## Problem\*

Goto Definitions still take take too much time to execute leaving user
with undesirable experience. On a bigger workspace (beyond simple test
cases, eg.
[noir-protocol-circuits](https://github.com/AztecProtocol/aztec-packages/tree/master/yarn-project/noir-protocol-circuits))
it can take as much as a second or more to perform goto request.

Resolves chore(lsp): improve goto performance by caching ast #3942

## Summary\*

Improvement from `~1s` down to `~11ms` on example project
[noir-protocol-circuits](https://github.com/AztecProtocol/aztec-packages/tree/master/yarn-project/noir-protocol-circuits)
is achieved by caching definitions resolved during file save or file
open.

The cost for this optimisation is when user opens noir source file, at
which point parsing and resolving is done. This however is not
noticeable for user as source code can be seen and operated on.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.