fix(docs): Fix docs build during releases (#3815)
Fixes the issue introduced in #3762, which caused the docs build step to
always attempt to fetch the last released version. Due to how
release-plase works, the last released version during a release is the
version being released, which is not available as a tag to be fetched.
To avoid this issue, we first check if the tag to be checked out is
available, and if not, just load the code snippets from master.