feat: add MVP `nargo export` command (#3870)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR adds the skeleton of a `nargo export` command which compiles all
of the functions in a library crate with the `#[export]` attribute into
an `export` directory. This is intended to be then fed into
`noir-codegen` in order to generate a TS library for executing these
functions.

## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: kevaundray <kevtheappdev@gmail.com>