feat(aztec.js)!: Move accounts out of aztec.js into new package (#3844)
Extracts account contract implementations and related classes, as well
as the default base classes, to a new `@aztec/accounts` package. Updates
docs to reference the new package, and includes it in the API ref.

This means that all scripts that relied on using accounts from the
Sandbox, creating new accounts, or using accounts already deployed, will
now need to import this package.

Since `accounts` is now needed along with `aztec.js` for most stuff,
including web, this PR also adds a new build step for end-to-end tests
to exercise web builds. Instead of building aztec.js via webpack as part
of aztec.js, we create a new "web application" as part of the end-to-end
tests, which includes both aztec.js and accounts, and test that. This
flow should be closer to how end-users will interact with aztec.js on
the web.

Fixes #3807