feat: Launch the monorepo in a codespace. (#3829)
This adds a `.devcontainer/devcontainer.json` file to the monorepo.
This allows one to launch a codespace terminal from github in browser.

The *expected* outcome of this would a codespace ready to go, for
actually developing our monorepo. A little bit like what our mainframe
offers. And I think it could well be worth configuring an image with all
the right tools to support this. But that wasn't really the goal of this
PR.

This actually just installs the latest master containers into a minimal
codespace, that allows a developer to unbox a project and run it against
a sandbox etc. It just makes it a bit easier to test this flow directly
in the monorepo.

Hey. I ❤️  monorepo.