fix: return error rather instead of panicking on invalid circuit (#3976)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR replaces a panic on an invalid circuit in ACVM JS with an error.
If this occurs it's likely due to serialisation differences rather than
a user passing in garbage bytes so I've added that to the error message.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.