fix: search all levels of struct nesting before codegenning primitive types (#3970)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

We previously had a bug that if a primitive type only exists below at
least 2 layers of nesting in the function arguments then we wouldn't
generate a definition for it.

We now search the entire depth of nested structs before generating the
primitive type definitions.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: kevaundray <kevtheappdev@gmail.com>