feat: support for u128 (#3913)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

First basic implementation of u128 in Noir, doing addition, subtraction,
multiplication and division, and with some conversion utilities.
It assumes the base field is larger than 128 bits.

## Additional Context



## Documentation\*

No documentation yet because it is an on-going development (more PRs
will follow) of an experimental feature

Check one:
- [X No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <tom@tomfren.ch>