test: end to end test node & pxe persistence (#3911)
This PR adds a new end-to-end test to check that database persistence is
working as expected for the node and PXE. This test suite uses the
'no-sandbox' docker-compose file in order to control the node's and
pxe's initialisation.

The test suite checks that four scenarios work correctly: new node, new
PXE, restored node, new PXE, new node, restored PXE and restored node,
restored PXE. All tests reuse the same L1 state and deployed account
contract and test contract. The only thing not tested is if the chain
advances while the node is shutdown. This will come in a separate PR.

---------

Co-authored-by: Santiago Palladino <santiago@aztecprotocol.com>