feat: remove truncation from brillig casts (#3997)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

As part of #3984, I noticed that I forgot to remove the truncation
behaviour from casts in brillig. This PR updates them to be a simple
`mov` instruction.

In order to do this I've added a proper implementation of
`Instruction::Truncate`.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.