chore: add test to disallow stdlib producing warnings (#4045)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

We've had a number of occasions where a feature is added which causes
the stdlib to produce warnings on every compilation. This isn't great so
we should catch and fix those warnings as we create them.

This PR adds a test which compiles a very simple program and asserts
that no warnings are omitted.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.