feat: Avoid unnecessary range checks by inspecting instructions for casts (#4039)
…

# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR removes unnecessary overflow checks in more scenarios than
previously. We do this by inspecting whether the inputs are
`Instruction::Cast`s which allows us to determine an more restrictive
upper bound on the value and reason about how this affects the result.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.