chore!: Ban nested slices (#4018)
# Description

## Problem\*

Resolves #4017 

## Summary\*

After this PR https://github.com/noir-lang/noir/pull/3979 and an
internal discussion we have decided to temporarily block nested slices.

This PR adds a check in the frontend and in the SSA against nested
slices. The check in the frontend makes sure any struct fields with a
nested slice (but without generics) are not blocked as well as any type
declarations with nested slices. In order to account for generics in
structs we also have a checked array codegen that makes sure we do not
have a nested slice.

## Additional Context

The actual nested slice code in the ACIR gen is somewhat intertwined so
I felt it would be best for a separate PR which will be a followup to
this one.

## Documentation\*

Check one:
- [] No documentation needed.
- [X] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.