chore: Remove nested slices SSA and ACIR codegen (#4021)
# Description

## Problem\*

Resolves #4019 

## Summary\*

This removes the relevant code introduced for nested slices. This may be
temporarily removed, but ultimately isn't too bad of a change as most of
the code is contained to the fill internal slices pass with some extra
logic in ACIR gen.

## Additional Context

See issue for extra context.

## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.