feat: bubble up `Instruction::Constrain`s to be applied as early as possible. (#4065)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR adds a very simple pass which moves any
`Instruction::Constrain`s to immediately after when all of their inputs
are available. This allows the ACIR gen pass to use this information
earlier as well as simplifying using these constraints in order to
inform constant folding in future.

## Additional Context

Pulled out from #4060 

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.