fix: hoist constraints on inputs to top of program (#4076)
# Description

## Problem\*

Followup to #4065 

## Summary\*

This PR fixes a small mistake in #4065. I mixed up the front and back of
the vector for the case where we don't find an instruction for which the
constraint's inputs are a result. This results in constraints on the
program inputs not being pushed up to the beginning of the program.

This PR fixes the index to insert these instructions so that we
constrain the inputs first.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.