chore: force programs to be fully rebuilt in integration tests (#4086)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

It's currently possible for an integration test to report true due to
the integration test using a cached artifact. This isn't an issue in CI
as no cached artifacts exist there but it can results in strange
behaviour locally.

This PR addresses this issue by forcing a full recompile.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.