refactor: archiver store (#3966)
This PR refactors the archiver store to use `@aztec/kv-store` instead of
lmdb directly. It is stacked on top of #4028.

The diff looks massive but all it does is split the archiver store into
individual pieces to better separate them conceptually (blocks, logs,
contracts, messages) and replace direct uses of lmdb with the data
structures from kv-store.