chore: improve warning message for returning a constant value (#4106)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This message is a bit strong for a warning (constant return values are
technically allowed, there's just generally not much reason for them).
It also doesn't give much information on why they may not be a good
idea.

I've updated the warning to something which I think is more useful.
Happy to take feedback on this however.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.