fix: check for tests in all packages before failing due to an unsatisfied test filter (#4114)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

Currently if you're running tests in a workspace and are filtering them
by name, if any of the packages doesn't contain a matching test then we
will immediately return an error. This PR fixes this so that we wait
until we have checked the entire workspace, should there still be no
matching tests then we throw an error but otherwise pass.


## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[Exceptional Case]** Documentation to be submitted in a separate
PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.