feat!: add opcode for poseidon2 permutation (#4214)
Related to issue: https://github.com/noir-lang/noir/issues/4037

The PR adds the opcode to ACIR and updates BB and Noir accordingly.
Furthermore you can use it via a foreign function in the stdlib. This
will generate the proper ACIR opcode but the solver will not be able to
solve it and BB will skip it.